aula 3

+++ server/src/routes/memories.ts
@@ -3,8 +3,15 @@ import { z } from "zod";
 import { prisma } from "./lib.prisma";
 
 export async function memoriesRoutes(app: FastifyInstance) {
-  app.get("/memories", async () => {
+  app.addHook("preHandler", async (request) => {
+    await request.jwtVerify();
+  });
+
+  app.get("/memories", async (request) => {
     const memories = await prisma.memory.findMany({
+      where: {
+        userId: request.user.sub,
+      },
       orderBy: {
         createdAt: "asc",
       },
@@ -18,7 +25,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
     });
   });
 
-  app.get("/memories/:id", async (request) => {
+  app.get("/memories/:id", async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     });
@@ -28,6 +35,12 @@ export async function memoriesRoutes(app: FastifyInstance) {
     const memory = await prisma.memory.findUniqueOrThrow({
       where: { id },
     });
+
+    if (!memory.isPublic && memory.userId !== request.user.sub) {
+      return reply.status(401).send({
+        message: "Você não tem permissão para ver está memória!",
+      });
+    }
 
     return memory;
   });
@@ -45,14 +58,14 @@ export async function memoriesRoutes(app: FastifyInstance) {
         content,
         coverUrl,
         isPublic,
-        userId: "016a58a0-568a-43b0-9078-cb45e0a46ced",
+        userId: request.user.sub,
       },
     });
 
     return memory;
   });
 
-  app.put("/memories/:id", async (request) => {
+  app.put("/memories/:id", async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     });
@@ -66,7 +79,19 @@ export async function memoriesRoutes(app: FastifyInstance) {
     });
     const { content, coverUrl, isPublic } = bodySchema.parse(request.body);
 
-    const memory = await prisma.memory.update({
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    });
+
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send({
+        message: "Você não tem permissão para atualizar está memória!",
+      });
+    }
+
+    memory = await prisma.memory.update({
       where: {
         id,
       },
@@ -80,17 +105,27 @@ export async function memoriesRoutes(app: FastifyInstance) {
     return memory;
   });
 
-  app.delete("/memories/:id", async (request) => {
+  app.delete("/memories/:id", async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     });
 
     const { id } = paramsSchema.parse(request.params);
 
-    const memory = await prisma.memory.delete({
-      where: { id },
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
     });
 
-    return memory;
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send({
+        message: "Você não tem permissão para atualizar está memória!",
+      });
+    }
+
+    await prisma.memory.delete({
+      where: { id },
+    });
   });
 }
